move setup_config to whole test suite
