+++ tests/server/test_config.py
@@ -6,8 +6,8 @@ from pathlib import Path
 
 from optimade.server.config import ServerConfig, CONFIG
 
-from .utils import SetClient
-from ..test_setup import setup_config
+from .utils import SetClient, setup_config
+
 
 setup_config()
