+++ tests/server/test_server_validation.py
@@ -4,8 +4,8 @@ import unittest
 
 from optimade.validator import ImplementationValidator
 
-from .utils import SetClient
-from ..test_setup import setup_config
+from .utils import SetClient, setup_config
+
 
 setup_config()
